requirements.txt
pyTelegramBotAPI==4.6.1
python-telegram-bot-calendar==1.0.5
schedule==1.1.0
gunicorn==20.1.0
boto3==1.26.15
botocore==1.29.15
python-dateutil==2.8.2